+++ Aka/CHANGELOG.md
@@ -96,3 +96,4 @@
   * `geoip <nick>` added
   * Bug fixes
   * `getconfig` to print the current settings for the network
+  * Spelling fix